# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/records": {
            "post": {
                "description": "insert new transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Insert record for wallet balance",
                "parameters": [
                    {
                        "description": "Models of record",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Record"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.BasicResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.BasicResponse"
                        }
                    }
                }
            }
        },
        "/records/history": {
            "post": {
                "description": "get history balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Get History from start to end date time",
                "parameters": [
                    {
                        "description": "Models of request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.History"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Record"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.BasicResponse"
                        }
                    }
                }
            }
        },
        "/records/latest": {
            "get": {
                "description": "get lattest balance",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wallets"
                ],
                "summary": "Get Latest Balance",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Record"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.BasicResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.BasicResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "models.History": {
            "type": "object",
            "required": [
                "endDatetime",
                "startDatetime"
            ],
            "properties": {
                "endDatetime": {
                    "type": "string"
                },
                "startDatetime": {
                    "type": "string"
                }
            }
        },
        "models.Record": {
            "type": "object",
            "required": [
                "amount",
                "datetime"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "datetime": {
                    "type": "string"
                }
            }
        }
    }
}